requirements.txt
gradio
pandas
planqk-service-sdk
plotly
requests
